HEADリクエストの削除

+++ pecaviewer/src/main/java/org/peercast/pecaviewer/chat/thumbnail/net/TooLargeFileException.kt
@@ -2,4 +2,4 @@ package org.peercast.pecaviewer.chat.thumbnail.net
 
 import java.io.IOException
 
-class TooLargeFileException(val size: Int) : IOException("large file: ${size / 1024}KB")
+class TooLargeFileException(val size: Long) : IOException("large file: ${size / 1024}KB")
